Fixed compiler warning

+++ harmony_model_checker/charm/graph.h
@@ -114,7 +114,7 @@ THIS SHOULD NOT BE DEFINED
     bool failed : 1;            // edge has failed (safety violation)
 };
 
-#define edge_sc(e)           ((struct step_condition *) (e->stc_id))
+#define edge_sc(e)           ((struct step_condition *) (uint64_t) (e->stc_id))
 // #define edge_sc(e)           (global.stc_table[(e)->stc_id])
 #define edge_input(e)        ((struct step_input *) &edge_sc(e)[1])
 #define edge_output(e)       (edge_sc(e)->u.completed)
